backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.25
mysql-connector-python==8.2.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
pydantic==2.5.0
email-validator==2.1.0
stripe==7.4.0
python-dotenv==1.0.0